Improvement update and delete user

+++ src/main/java/com/ut7pro/ut7pro/services/UserService.java
@@ -8,6 +8,7 @@ import com.ut7pro.ut7pro.exceptions.NotFoundException;
 import com.ut7pro.ut7pro.models.User;
 import com.ut7pro.ut7pro.repositories.UserRepository;
 import java.util.List;
+import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -23,18 +24,38 @@ public class UserService {
 
     public User getUserById(Long id) {
         return userRepository.findById(id)
-                .orElseThrow(() -> new NotFoundException("User not found with ID: "+ id));
+                .orElseThrow(() -> new NotFoundException("User not found with ID: " + id));
     }
 
     public User createUser(User newUser) {
         return userRepository.save(newUser);
     }
 
-    public User updateUser(User updatedUser) {
-        return userRepository.save(updatedUser);
+    public User updateUser(User updatedUser, Long id) {
+        // Verificar si el usuario existe en la base de datos
+        Optional<User> optionalUser = userRepository.findById(id);
+        if (optionalUser.isEmpty()) {
+            throw new NotFoundException("User not found with ID: " + id);
+        }
+        // Obtener el usuario existente
+        User existingUser = optionalUser.get();
+
+        // Aplicar las actualizaciones en el usuario existente
+        existingUser.setName(updatedUser.getName());
+        existingUser.setEmail(updatedUser.getEmail());
+        existingUser.setToken(updatedUser.getToken());
+        existingUser.setRolId(updatedUser.getRolId());
+
+        // Guardar los cambios en la base de datos
+        return userRepository.save(existingUser);
     }
 
     public void deleteUser(Long id) {
+        // Verificar si el usuario existe en la base de datos
+        Optional<User> optionalUser = userRepository.findById(id);
+        if (optionalUser.isEmpty()) {
+            throw new NotFoundException("User not found with ID: " + id);
+        }
         userRepository.deleteById(id);
     }
 }
